More updates for the new naming schema. Firehose.Batches, Kinesis.Records, SQS.Messages, SNS.Notifications, S3.Notifications
