+++ examples/lib/stream-processing.ts
@@ -41,7 +41,6 @@ const enrichments = new DynamoDB.Table(stack, 'Enrichments', {
   billingMode: BillingMode.PAY_PER_REQUEST
 });
 
-
 /**
  * Schedule a Lambda Function to send a (dummy) message to the SNS topic:
  * 
@@ -91,7 +90,7 @@ new Lambda.ExecutorService().schedule(stack, 'DummyData', {
  *
  * SNS -> Lambda
  */
-topic.stream().forEach(stack, 'ForEachNotification', {
+topic.notifications().forEach(stack, 'ForEachNotification', {
   async handle(message) {
     console.log(`received notification '${message.key}' with a delay of ${new Date().getTime() - message.timestamp.getTime()}ms`);
   }
@@ -112,7 +111,7 @@ const queue = topic.toSQSQueue(stack, 'Queue');
  *                v
  * SQS Queue -> Lambda -> Kinesis Stream
  */
-const stream = queue.stream() // stream gives us a nice chainable API for resources like queues, streams, topics etc.
+const stream = queue.messages() // gives us a nice chainable API
   .map({
     depends: enrichments.readAccess(),
     handle: async(message, e) => {
@@ -154,7 +153,7 @@ const database = new glue.Database(stack, 'Database', {
   databaseName: 'my_database'
 });
 stream
-  .toFirehoseDeliveryStream(stack, 'ToS3').stream()
+  .toFirehoseDeliveryStream(stack, 'ToS3').batches()
   .toGlueTable(stack, 'ToGlue', {
     database,
     tableName: 'my_table',
