+++ packages/punchcard/lib/firehose/index.ts
@@ -1,4 +1,5 @@
+export * from './batches';
+export * from './client';
 export * from './collector';
 export * from './delivery-stream';
 export * from './event';
-export * from './stream';
