+++ packages/punchcard/lib/s3/index.ts
@@ -1,2 +1,4 @@
 export * from './bucket';
+export * from './client';
 export * from './event';
+export * from './notifications';
